making changes according to Brandon's feedback
